refactor(igx-expansion-panel): Remove unnecessary samples
