+++ src/app/layouts/expansion-panel/expansion-sample-4/expansion-sample-4.component.html
@@ -1,31 +1,29 @@
-<div class="sample-wrapper">
-    <igx-expansion-panel [animationSettings]="animationSettings" (onCollapsed)="handleCollapsed($event)"
-        (onExpanded)="handleExpanded($event)" #collapsibleComponent>
-        <igx-expansion-panel-header [iconPosition]="iconPosition" (onInteraction)="onInteraction($event)">
-            <igx-expansion-panel-title>Current Winning Player:</igx-expansion-panel-title>
-            <igx-expansion-panel-description>{{ getWinningPlayer }}</igx-expansion-panel-description>
-            <button igxButton class='igx-button--flat' (click)="resetScore($event)">Reset score</button>
-            <igx-expansion-panel-icon *ngIf="templatedIcon">
-                <!-- <igx-icon fontSet="material">{{collapsed() ? 'arrow_drop_down':'arrow_drop_up'}}</igx-icon> -->
-                {{collapsed() ? 'Expand':'Collapse'}}
-            </igx-expansion-panel-icon>
+<div class="sample-holder sample-center--col">
+    <igx-expansion-panel class="my-expansion-panel">
+        <igx-expansion-panel-header (onInteraction)="onInteraction($event)" [disabled]="false">
+            <igx-expansion-panel-title class="sample-title">HTML5</igx-expansion-panel-title>
         </igx-expansion-panel-header>
         <igx-expansion-panel-body>
-            <igx-grid [data]="data">
-                <igx-column [sortable]="true" [field]="'Game'" [header]="'Game'" [width]="'120px'" [pinned]="true">
-                    <ng-template igxCell let-value>
-                        <span class="game-board__grid__match">{{value}}</span>
-                    </ng-template>
-                </igx-column>
-                <igx-column [sortable]="true" [field]="'Player 1'" [header]="'Score: Player 1'">
-                </igx-column>
-                <igx-column [sortable]="true" [field]="'Player 2'" [header]="'Score: Player 2'">
-                </igx-column>
-            </igx-grid>
+            <div class="example-content">
+                HTML5 is a software solution stack that defines the properties and behaviors of web page content by implementing a markup-based pattern to it.            </div>
         </igx-expansion-panel-body>
     </igx-expansion-panel>
-    <div class="buttons-panel">
-        <button igxButton class="igx-button--raised update-button" (click)="templateIcon()">{{ templatedIcon ? 'Disabled' : 'Enable'}} icon template</button>
-        <button igxButton class="igx-button--raised update-button" (click)="toggleLeftRight()">Current Position: {{ iconPosition }}</button>
-    </div>
+    <igx-expansion-panel class="my-expansion-panel">
+        <igx-expansion-panel-header (onInteraction)="onInteraction($event)" [disabled]="false">
+            <igx-expansion-panel-title class="sample-title">CSS3</igx-expansion-panel-title>
+        </igx-expansion-panel-header>
+        <igx-expansion-panel-body>
+            <div class="example-content">
+                Cascading Style Sheets (CSS) is a style sheet language used for describing the presentation of a document written in a markup language like HTML            </div>
+        </igx-expansion-panel-body>
+    </igx-expansion-panel>
+    <igx-expansion-panel class="my-expansion-panel">
+        <igx-expansion-panel-header (onInteraction)="onInteraction($event)" [disabled]="false">
+            <igx-expansion-panel-title class="sample-title">SASS/SCSS</igx-expansion-panel-title>
+        </igx-expansion-panel-header>
+        <igx-expansion-panel-body>
+            <div class="example-content">
+                Sass is a preprocessor scripting language that is interpreted or compiled into Cascading Style Sheets (CSS).            </div>
+        </igx-expansion-panel-body>
+    </igx-expansion-panel>
 </div>
