+++ src/app/layouts/expansion-panel/expansion-sample-5/expansion-sample-5.component.html
@@ -1,108 +1,10 @@
-<div class="sample-holder sample-center--col">
-    <igx-expansion-panel class="my-expansion-panel">
-        <igx-expansion-panel-header (onInteraction)="onInteraction($event)" [disabled]="false">
-            <igx-expansion-panel-title class="sample-title">Personal Information</igx-expansion-panel-title>
+<div class="sample-holder">
+    <igx-expansion-panel [animationSettings] = "animationSettingsCustom" class="my-expansion-panel">
+        <igx-expansion-panel-header>
+            <igx-expansion-panel-title class="sample-title">Angular</igx-expansion-panel-title>
         </igx-expansion-panel-header>
-        <igx-expansion-panel-body>
-            <div class="body">
-                <igx-input-group class="group">
-                    <input igxInput name="fullName" type="text" [(ngModel)]="user.fullName"/>
-                    <label igxLabel for="fullName">Full Name</label>
-                    <igx-suffix>
-                        <igx-icon>person</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <igx-prefix>+359</igx-prefix>
-                    <label igxLabel for="phone">Phone</label>
-                    <input type="number" igxInput name="phone" [(ngModel)]="user.phone" />
-                    <igx-suffix>
-                        <igx-icon>phone</igx-icon>
-                    </igx-suffix>
-                    <igx-hint position="start">Ex.: +359 555 123 456</igx-hint>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="email">Email address</label>
-                    <input igxInput name="email" type="email" [(ngModel)]="user.email"/>
-                    <igx-suffix>
-                        <igx-icon>email</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-            </div>
-        </igx-expansion-panel-body>
-    </igx-expansion-panel>
-    <igx-expansion-panel class="my-expansion-panel">
-        <igx-expansion-panel-header (onInteraction)="onInteraction($event)" [disabled]="false">
-            <igx-expansion-panel-title class="sample-title">Billing Address</igx-expansion-panel-title>
-        </igx-expansion-panel-header>
-        <igx-expansion-panel-body>
-            <div class="body">
-                <igx-input-group class="group">
-                    <input igxInput name="address" type="text" [(ngModel)]="shippingAddress.address"/>
-                    <label igxLabel for="address">Billing Address:</label>
-                    <igx-suffix>
-                        <igx-icon>add_location</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="city">City:</label>
-                    <input igxInput name="city" type="text" [(ngModel)]="shippingAddress.city"/>
-                    <igx-suffix>
-                        <igx-icon>location_city</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="state">State:</label>
-                    <input igxInput name="state" type="text" [(ngModel)]="shippingAddress.state"/>
-                    <igx-suffix>
-                        <igx-icon>terrain</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="zipCode">Zip Code:</label>
-                    <input igxInput name="zipCode" type="text" [(ngModel)]="shippingAddress.zipCode">
-                    <igx-suffix>
-                        <igx-icon>mail_outline</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-            </div>
-        </igx-expansion-panel-body>
-    </igx-expansion-panel>
-    <igx-expansion-panel class="my-expansion-panel">
-        <igx-expansion-panel-header (onInteraction)="onInteraction($event)" [disabled]="false">
-            <igx-expansion-panel-title class="sample-title">Shipping Address</igx-expansion-panel-title>
-        </igx-expansion-panel-header>
-        <igx-expansion-panel-body>
-            <div class="body">
-                <igx-input-group class="group">
-                    <input igxInput name="address" type="text" [(ngModel)]="billingAddress.address"/>
-                    <label igxLabel for="address">Shipping Address:</label>
-                    <igx-suffix>
-                        <igx-icon>add_location</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="city">City:</label>
-                    <input igxInput name="city" type="text" [(ngModel)]="billingAddress.city"/>
-                    <igx-suffix>
-                        <igx-icon>location_city</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="state">State:</label>
-                    <input igxInput name="state" type="text" [(ngModel)]="billingAddress.state"/>
-                    <igx-suffix>
-                        <igx-icon>terrain</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-                <igx-input-group class="group">
-                    <label igxLabel for="zipCode">Zip Code:</label>
-                    <input igxInput name="zipCode" type="text" [(ngModel)]="billingAddress.zipCode">
-                    <igx-suffix>
-                        <igx-icon>mail_outline</igx-icon>
-                    </igx-suffix>
-                </igx-input-group>
-            </div>
+        <igx-expansion-panel-body class="example-content">
+            Angular (commonly referred to as "Angular 2+" or "Angular v2 and above") is a TypeScript-based open-source web application framework led by the Angular Team at Google and by a community of individuals and corporations.
         </igx-expansion-panel-body>
     </igx-expansion-panel>
 </div>
