+++ src/app/layouts/expansion-panel/expansion-sample-5/expansion-sample-5.component.scss
@@ -1,27 +1,17 @@
-$ep-sample-border: 1px solid rgba(174, 174, 174, 0.25);
-
-.sample-holder{
+.sample-holder {
     display: flex;
-    flex-direction: column;
     align-items: flex-start;
     justify-content: flex-start;
     height: 100%;
-    padding: 16px;
 }
 
 .my-expansion-panel {
     max-width: 500px;
     min-width: 300px;
     width: 100%;
-    border-top: $ep-sample-border;
-    border-left: $ep-sample-border;
-    border-right: $ep-sample-border;
-
-    &:last-of-type {
-        border-bottom: $ep-sample-border;
-    }
+    border: 1px solid rgba(171, 171, 171, 0.3)
 }
 
-.body {
-    padding: 0 16px 16px 16px;
+.example-content {
+    padding: 16px;
 }
