--- src/app/layouts/expansion-panel/expansion-sample-6/expansion-sample-6.component.html
@@ -1,32 +1,0 @@
-<div class="sample-wrapper">
-    <igx-expansion-panel [animationSettings] = "animationSettingsCustom" class="content__collapsible">
-        <igx-expansion-panel-header>
-            <igx-expansion-panel-title class="sample-title">Personal Information</igx-expansion-panel-title>
-        </igx-expansion-panel-header>
-        <igx-expansion-panel-body class="body">
-            <igx-input-group class="group">
-                <label igxLabel for="fullName">Full Name</label>
-                <input igxInput autocomplete="off" type="text" name="fullName" [(ngModel)]="user.fullName"/>
-                <igx-suffix>
-                    <igx-icon>person</igx-icon>
-                </igx-suffix>
-            </igx-input-group>
-            <igx-input-group class="group">
-                <igx-prefix>+359</igx-prefix>
-                <label igxLabel for="phone">Phone</label>
-                <input igxInput autocomplete="off" type="number" name="phone" [(ngModel)]="user.phone" />
-                <igx-suffix>
-                    <igx-icon>phone</igx-icon>
-                </igx-suffix>
-                <igx-hint position="start">Ex.: +359 555 123 456</igx-hint>
-            </igx-input-group>
-            <igx-input-group class="group">
-                <label igxLabel for="email">Email address</label>
-                <input igxInput autocomplete="off" type="email" name="email" [(ngModel)]="user.email"/>
-                <igx-suffix>
-                    <igx-icon>email</igx-icon>
-                </igx-suffix>
-            </igx-input-group>
-        </igx-expansion-panel-body>
-    </igx-expansion-panel>
-</div>
