--- src/app/layouts/expansion-panel/expansion-sample-6/expansion-sample-6.component.scss
@@ -1,18 +1,0 @@
-.sample-wrapper {
-    width: 400px;
-}
-
-.content__collapsible {
-    width: 100%;
-    border: 1px solid hsl(0, 0%, 82%);
-    border-radius: .3em;
-    box-shadow: 0 1px 2px hsl(0, 0%, 82%);
-}
-
-igx-expansion-panel-body {
-    padding: 0em 1em 1em 1em;
-}
-
-.sample-title {
-    font-weight: 900;
-}
