--- src/app/layouts/expansion-panel/expansion-sample-7/expansion-sample-7.component.ts
@@ -1,19 +1,0 @@
-import { Component, ViewChild } from "@angular/core";
-import { IgxExpansionPanelComponent } from "igniteui-angular";
-import { data as weatherData } from "./weather-data";
-@Component({
-    // tslint:disable-next-line:component-selector
-    selector: "app-expansion-sample-7",
-    styleUrls: ["./expansion-sample-7.component.scss"],
-    templateUrl: "./expansion-sample-7.component.html"
-})
-export class ExpansionPanelSample7Component {
-
-    @ViewChild(IgxExpansionPanelComponent, { static: true })
-    public panel: IgxExpansionPanelComponent;
-    public data = weatherData;
-
-    public toggleDetails() {
-        this.panel.toggle();
-    }
-}
