--- src/app/layouts/expansion-panel/expansion-sample-7/weather-data.ts
@@ -1,58 +1,0 @@
-// tslint:disable:object-literal-sort-keys
-export const data = {
-  city: "Sofia",
-  humidity: "44%",
-  precipitation: "5%",
-  windSpeed: 279,
-  dateTime: "10/7/2018, 14:35:00 PM",
-  today: {
-    name: "Sunday",
-    tempMax: 25,
-    tempMin: 15,
-    description: "Sunday",
-    iconColor: "gray",
-    iconName: "fa-cloud"
-  },
-  daysOfWeek: [
-    {
-      name: "Monday",
-      tempMax: 22,
-      tempMin: 15,
-      description: "Sunny",
-      iconColor: "orange",
-      iconName: "fa-sun"
-    },
-    {
-      name: "Tuesday",
-      tempMax: 25,
-      tempMin: 14,
-      description: "Rainy",
-      iconColor: "blue",
-      iconName: "fa-tint"
-    },
-    {
-      name: "Wednesday",
-      tempMax: 28,
-      tempMin: 18,
-      description: "Sunny",
-      iconColor: "orange",
-      iconName: "fa-sun"
-    },
-    {
-      name: "Thursday",
-      tempMax: 20,
-      tempMin: 12,
-      description: "Few Clouds",
-      iconColor: "gray",
-      iconName: "fa-cloud"
-    },
-    {
-      name: "Friday",
-      tempMax: 23,
-      tempMin: 13,
-      description: "Light Rain",
-      iconColor: "blue",
-      iconName: "fa-tint"
-    }
-  ]
-};
